# lib/api-zod/src/generated/types/assignment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EduWeb - Platform for teachers to assign homework and students to submit
 * OpenAPI spec version: 0.1.0
 */
import type { Submission } from './submission';

export interface Assignment {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  classId: number;
  className: string;
  /** @nullable */
  dueDate?: string | null;
  submissionCount?: number;
  mySubmission?: Submission;
  createdAt: string;
}