# lib/api-zod/src/generated/types/assignmentDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EduWeb - Platform for teachers to assign homework and students to submit
 * OpenAPI spec version: 0.1.0
 */
import type { Submission } from './submission';

export interface AssignmentDetail {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  classId: number;
  className: string;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
  submissions?: Submission[];
}